Monoid Sum keeps its name and wire id; SQL-semantics aggregate is StandardSum, used only by list_sum
