+++ vortex-array/benches/aggregate_grouped.rs
@@ -19,8 +19,8 @@ use vortex_array::aggregate_fn::DynGroupedAccumulator;
 use vortex_array::aggregate_fn::GroupedAccumulator;
 use vortex_array::aggregate_fn::NumericalAggregateOpts;
 use vortex_array::aggregate_fn::fns::count::Count;
+use vortex_array::aggregate_fn::fns::standard_sum::StandardSum;
 use vortex_array::aggregate_fn::fns::sum::Sum;
-use vortex_array::aggregate_fn::fns::total::Total;
 use vortex_array::arrays::ListViewArray;
 use vortex_array::arrays::PrimitiveArray;
 use vortex_array::arrays::VarBinViewArray;
@@ -180,7 +180,7 @@ fn sum_i32_nullable_all_valid(bencher: Bencher) {
     let input = i32_nullable_all_valid_input();
     bencher
         .with_inputs(|| &input)
-        .bench_refs(|input| grouped_accumulator(input, Total));
+        .bench_refs(|input| grouped_accumulator(input, Sum));
 }
 
 #[divan::bench]
@@ -188,7 +188,7 @@ fn sum_i32_clustered_nulls(bencher: Bencher) {
     let input = i32_clustered_nulls_input();
     bencher
         .with_inputs(|| &input)
-        .bench_refs(|input| grouped_accumulator(input, Total));
+        .bench_refs(|input| grouped_accumulator(input, Sum));
 }
 
 #[divan::bench]
@@ -196,7 +196,7 @@ fn sum_f64_all_valid(bencher: Bencher) {
     let input = f64_all_valid_input();
     bencher
         .with_inputs(|| &input)
-        .bench_refs(|input| grouped_accumulator(input, Total));
+        .bench_refs(|input| grouped_accumulator(input, Sum));
 }
 
 #[divan::bench]
@@ -204,39 +204,39 @@ fn sum_f64_clustered_nulls(bencher: Bencher) {
     let input = f64_clustered_nulls_input();
     bencher
         .with_inputs(|| &input)
-        .bench_refs(|input| grouped_accumulator(input, Total));
+        .bench_refs(|input| grouped_accumulator(input, Sum));
 }
 
 #[divan::bench]
-fn sql_sum_i32_nullable_all_valid(bencher: Bencher) {
+fn standard_sum_i32_nullable_all_valid(bencher: Bencher) {
     let input = i32_nullable_all_valid_input();
     bencher
         .with_inputs(|| &input)
-        .bench_refs(|input| grouped_accumulator(input, Sum));
+        .bench_refs(|input| grouped_accumulator(input, StandardSum));
 }
 
 #[divan::bench]
-fn sql_sum_i32_clustered_nulls(bencher: Bencher) {
+fn standard_sum_i32_clustered_nulls(bencher: Bencher) {
     let input = i32_clustered_nulls_input();
     bencher
         .with_inputs(|| &input)
-        .bench_refs(|input| grouped_accumulator(input, Sum));
+        .bench_refs(|input| grouped_accumulator(input, StandardSum));
 }
 
 #[divan::bench]
-fn sql_sum_f64_all_valid(bencher: Bencher) {
+fn standard_sum_f64_all_valid(bencher: Bencher) {
     let input = f64_all_valid_input();
     bencher
         .with_inputs(|| &input)
-        .bench_refs(|input| grouped_accumulator(input, Sum));
+        .bench_refs(|input| grouped_accumulator(input, StandardSum));
 }
 
 #[divan::bench]
-fn sql_sum_f64_clustered_nulls(bencher: Bencher) {
+fn standard_sum_f64_clustered_nulls(bencher: Bencher) {
     let input = f64_clustered_nulls_input();
     bencher
         .with_inputs(|| &input)
-        .bench_refs(|input| grouped_accumulator(input, Sum));
+        .bench_refs(|input| grouped_accumulator(input, StandardSum));
 }
 
 #[divan::bench]
