+++ vortex-array/src/aggregate_fn/fns/standard_sum/primitive.rs
@@ -26,12 +26,19 @@ pub(super) fn accumulate_primitive(
     p: &PrimitiveArray,
     ctx: &mut ExecutionCtx,
     skip_nans: bool,
+    seen: &mut bool,
 ) -> VortexResult<bool> {
     let mask = p.as_ref().validity()?.execute_mask(p.as_ref().len(), ctx)?;
     match mask.slices() {
         AllOr::None => Ok(false),
-        AllOr::All => accumulate_primitive_all(inner, p, skip_nans),
-        AllOr::Some(slices) => accumulate_primitive_valid(inner, p, slices, skip_nans),
+        AllOr::All => {
+            *seen |= !p.as_ref().is_empty();
+            accumulate_primitive_all(inner, p, skip_nans)
+        }
+        AllOr::Some(slices) => {
+            *seen |= !slices.is_empty();
+            accumulate_primitive_valid(inner, p, slices, skip_nans)
+        }
     }
 }
 
@@ -63,7 +70,7 @@ fn accumulate_primitive_all(
     }
 }
 
-/// Total the values of a float slice into an `f64` accumulator. When `skip_nans` is set, NaN values
+/// StandardSum the values of a float slice into an `f64` accumulator. When `skip_nans` is set, NaN values
 /// are skipped to match the scalar `sum` semantics; otherwise any NaN poisons the accumulator to
 /// NaN. Floats cannot overflow the accumulator, so this never reports saturation.
 pub(super) fn sum_float_all<T: NativePType>(acc: &mut f64, slice: &[T], skip_nans: bool) {
@@ -80,7 +87,7 @@ pub(super) fn sum_float_all<T: NativePType>(acc: &mut f64, slice: &[T], skip_nan
     }
 }
 
-/// Total all values into a `u64` accumulator. For types narrower than 64 bits, values are summed in
+/// StandardSum all values into a `u64` accumulator. For types narrower than 64 bits, values are summed in
 /// chunks of [`SUM_CHUNK`] with a single checked add per chunk, which lets the inner loop vectorize
 /// to packed widening adds. `u64` input keeps a per-element checked add since a chunk of `u64`s
 /// could itself overflow. Returns `true` on overflow.
@@ -127,7 +134,7 @@ where
     false
 }
 
-/// Total the valid elements, described as contiguous `[start, end)` runs of set validity bits. Each
+/// StandardSum the valid elements, described as contiguous `[start, end)` runs of set validity bits. Each
 /// run is a slice of fully-valid values, so it reuses the same vectorized reduction as the
 /// all-valid path instead of a per-element validity branch.
 fn accumulate_primitive_valid(
@@ -188,8 +195,8 @@ mod tests {
     use crate::aggregate_fn::Accumulator;
     use crate::aggregate_fn::DynAccumulator;
     use crate::aggregate_fn::NumericalAggregateOpts;
-    use crate::aggregate_fn::fns::total::Total;
-    use crate::aggregate_fn::fns::total::total;
+    use crate::aggregate_fn::fns::standard_sum::StandardSum;
+    use crate::aggregate_fn::fns::standard_sum::standard_sum;
     use crate::array_session;
     use crate::arrays::ConstantArray;
     use crate::arrays::PrimitiveArray;
@@ -206,7 +213,7 @@ mod tests {
     #[test]
     fn sum_i32() -> VortexResult<()> {
         let arr = PrimitiveArray::new(buffer![1i32, 2, 3, 4], Validity::NonNullable).into_array();
-        let result = total(&arr, &mut array_session().create_execution_ctx())?;
+        let result = standard_sum(&arr, &mut array_session().create_execution_ctx())?;
         assert_eq!(result.as_primitive().typed_value::<i64>(), Some(10));
         Ok(())
     }
@@ -214,7 +221,7 @@ mod tests {
     #[test]
     fn sum_u8() -> VortexResult<()> {
         let arr = PrimitiveArray::new(buffer![10u8, 20, 30], Validity::NonNullable).into_array();
-        let result = total(&arr, &mut array_session().create_execution_ctx())?;
+        let result = standard_sum(&arr, &mut array_session().create_execution_ctx())?;
         assert_eq!(result.as_primitive().typed_value::<u64>(), Some(60));
         Ok(())
     }
@@ -223,7 +230,7 @@ mod tests {
     fn sum_f64() -> VortexResult<()> {
         let arr =
             PrimitiveArray::new(buffer![1.5f64, 2.5, 3.0], Validity::NonNullable).into_array();
-        let result = total(&arr, &mut array_session().create_execution_ctx())?;
+        let result = standard_sum(&arr, &mut array_session().create_execution_ctx())?;
         assert_eq!(result.as_primitive().typed_value::<f64>(), Some(7.0));
         Ok(())
     }
@@ -231,7 +238,7 @@ mod tests {
     #[test]
     fn sum_with_nulls() -> VortexResult<()> {
         let arr = PrimitiveArray::from_option_iter([Some(2i32), None, Some(4)]).into_array();
-        let result = total(&arr, &mut array_session().create_execution_ctx())?;
+        let result = standard_sum(&arr, &mut array_session().create_execution_ctx())?;
         assert_eq!(result.as_primitive().typed_value::<i64>(), Some(6));
         Ok(())
     }
@@ -251,7 +258,7 @@ mod tests {
             Some(6),
         ])
         .into_array();
-        let result = total(&arr, &mut array_session().create_execution_ctx())?;
+        let result = standard_sum(&arr, &mut array_session().create_execution_ctx())?;
         assert_eq!(result.as_primitive().typed_value::<i64>(), Some(21));
         Ok(())
     }
@@ -259,23 +266,24 @@ mod tests {
     #[test]
     fn sum_all_null() -> VortexResult<()> {
         let arr = PrimitiveArray::from_option_iter([None::<i32>, None, None]).into_array();
-        let result = total(&arr, &mut array_session().create_execution_ctx())?;
-        assert_eq!(result.as_primitive().typed_value::<i64>(), Some(0));
+        let result = standard_sum(&arr, &mut array_session().create_execution_ctx())?;
+        // SQL `SUM`: no valid values yields null.
+        assert!(result.is_null());
         Ok(())
     }
 
     #[test]
     fn sum_all_invalid_float() -> VortexResult<()> {
         let arr = PrimitiveArray::from_option_iter::<f32, _>([None, None, None]).into_array();
-        let result = total(&arr, &mut array_session().create_execution_ctx())?;
-        assert_eq!(result, Scalar::primitive(0f64, Nullable));
+        let result = standard_sum(&arr, &mut array_session().create_execution_ctx())?;
+        assert_eq!(result, Scalar::null(DType::Primitive(PType::F64, Nullable)));
         Ok(())
     }
 
     #[test]
     fn sum_buffer_i32() -> VortexResult<()> {
         let arr = buffer![1, 1, 1, 1].into_array();
-        let result = total(&arr, &mut array_session().create_execution_ctx())?;
+        let result = standard_sum(&arr, &mut array_session().create_execution_ctx())?;
         assert_eq!(result.as_primitive().as_::<i32>(), Some(4));
         Ok(())
     }
@@ -283,26 +291,27 @@ mod tests {
     #[test]
     fn sum_buffer_f64() -> VortexResult<()> {
         let arr = buffer![1., 1., 1., 1.].into_array();
-        let result = total(&arr, &mut array_session().create_execution_ctx())?;
+        let result = standard_sum(&arr, &mut array_session().create_execution_ctx())?;
         assert_eq!(result.as_primitive().as_::<f32>(), Some(4.));
         Ok(())
     }
 
     #[test]
-    fn sum_empty_produces_zero() -> VortexResult<()> {
+    fn sum_empty_is_null() -> VortexResult<()> {
         let dtype = DType::Primitive(PType::I32, Nullability::NonNullable);
-        let mut acc = Accumulator::try_new(Total, NumericalAggregateOpts::default(), dtype)?;
+        let mut acc = Accumulator::try_new(StandardSum, NumericalAggregateOpts::default(), dtype)?;
         let result = acc.finish()?;
-        assert_eq!(result.as_primitive().typed_value::<i64>(), Some(0));
+        // SQL `SUM`: the sum over no values is null.
+        assert!(result.is_null());
         Ok(())
     }
 
     #[test]
-    fn sum_empty_f64_produces_zero() -> VortexResult<()> {
+    fn sum_empty_f64_is_null() -> VortexResult<()> {
         let dtype = DType::Primitive(PType::F64, Nullability::NonNullable);
-        let mut acc = Accumulator::try_new(Total, NumericalAggregateOpts::default(), dtype)?;
+        let mut acc = Accumulator::try_new(StandardSum, NumericalAggregateOpts::default(), dtype)?;
         let result = acc.finish()?;
-        assert_eq!(result.as_primitive().typed_value::<f64>(), Some(0.0));
+        assert!(result.is_null());
         Ok(())
     }
 
@@ -313,7 +322,7 @@ mod tests {
             Validity::NonNullable,
         )
         .into_array();
-        let result = total(&arr, &mut array_session().create_execution_ctx())?;
+        let result = standard_sum(&arr, &mut array_session().create_execution_ctx())?;
         assert_eq!(result.as_primitive().typed_value::<f64>(), Some(6.0));
         Ok(())
     }
@@ -322,7 +331,7 @@ mod tests {
     fn sum_f32_with_nan() -> VortexResult<()> {
         let arr =
             PrimitiveArray::new(buffer![1.0f32, f32::NAN, 4.0], Validity::NonNullable).into_array();
-        let result = total(&arr, &mut array_session().create_execution_ctx())?;
+        let result = standard_sum(&arr, &mut array_session().create_execution_ctx())?;
         assert_eq!(result.as_primitive().typed_value::<f64>(), Some(5.0));
         Ok(())
     }
@@ -331,7 +340,7 @@ mod tests {
     fn sum_f64_with_nan_and_nulls() -> VortexResult<()> {
         let arr = PrimitiveArray::from_option_iter([Some(1.0f64), None, Some(f64::NAN), Some(3.0)])
             .into_array();
-        let result = total(&arr, &mut array_session().create_execution_ctx())?;
+        let result = standard_sum(&arr, &mut array_session().create_execution_ctx())?;
         assert_eq!(result.as_primitive().typed_value::<f64>(), Some(4.0));
         Ok(())
     }
@@ -340,17 +349,17 @@ mod tests {
     fn sum_all_nan() -> VortexResult<()> {
         let arr =
             PrimitiveArray::new(buffer![f64::NAN, f64::NAN], Validity::NonNullable).into_array();
-        let result = total(&arr, &mut array_session().create_execution_ctx())?;
+        let result = standard_sum(&arr, &mut array_session().create_execution_ctx())?;
         assert_eq!(result.as_primitive().typed_value::<f64>(), Some(0.0));
         Ok(())
     }
 
-    /// Total an array with explicit [`NumericalAggregateOpts`] (test-only helper).
+    /// StandardSum an array with explicit [`NumericalAggregateOpts`] (test-only helper).
     fn sum_with_options(
         arr: &crate::ArrayRef,
         options: NumericalAggregateOpts,
     ) -> VortexResult<Scalar> {
-        let mut acc = Accumulator::try_new(Total, options, arr.dtype().clone())?;
+        let mut acc = Accumulator::try_new(StandardSum, options, arr.dtype().clone())?;
         acc.accumulate(arr, &mut array_session().create_execution_ctx())?;
         acc.finish()
     }
@@ -388,13 +397,15 @@ mod tests {
 
     #[test]
     fn sum_not_skipping_uses_cached_sum_when_nan_free() -> VortexResult<()> {
-        // With an exact NaNCount of zero, the planted exact Total stat is usable as-is.
+        // With an exact NaNCount of zero, the planted exact StandardSum stat is usable as-is.
         let arr =
             PrimitiveArray::new(buffer![1.0f64, 2.0, 3.0], Validity::NonNullable).into_array();
         arr.statistics()
             .set(Stat::NaNCount, Precision::Exact(ScalarValue::from(0u64)));
         arr.statistics()
             .set(Stat::Sum, Precision::Exact(ScalarValue::from(42.0f64)));
+        arr.statistics()
+            .set(Stat::NullCount, Precision::Exact(ScalarValue::from(0u64)));
         let result = sum_with_options(&arr, NumericalAggregateOpts::include_nans())?;
         assert_eq!(result.as_primitive().typed_value::<f64>(), Some(42.0));
         Ok(())
@@ -419,12 +430,12 @@ mod tests {
             Validity::NonNullable,
         )
         .into_array();
-        let acc = total(&batch, &mut array_session().create_execution_ctx())?;
+        let acc = standard_sum(&batch, &mut array_session().create_execution_ctx())?;
         // INFINITY + NEG_INFINITY = NaN, which is treated as saturated
         assert!(acc.as_primitive().typed_value::<f64>().unwrap().is_nan());
 
         let mut acc = Accumulator::try_new(
-            Total,
+            StandardSum,
             NumericalAggregateOpts::default(),
             DType::Primitive(PType::F64, Nullability::NonNullable),
         )?;
@@ -436,7 +447,7 @@ mod tests {
     #[test]
     fn sum_checked_overflow() -> VortexResult<()> {
         let arr = PrimitiveArray::new(buffer![i64::MAX, 1i64], Validity::NonNullable).into_array();
-        let result = total(&arr, &mut array_session().create_execution_ctx())?;
+        let result = standard_sum(&arr, &mut array_session().create_execution_ctx())?;
         assert!(result.is_null());
         Ok(())
     }
@@ -444,7 +455,7 @@ mod tests {
     #[test]
     fn sum_checked_overflow_is_saturated() -> VortexResult<()> {
         let dtype = DType::Primitive(PType::I64, Nullability::NonNullable);
-        let mut acc = Accumulator::try_new(Total, NumericalAggregateOpts::default(), dtype)?;
+        let mut acc = Accumulator::try_new(StandardSum, NumericalAggregateOpts::default(), dtype)?;
         assert!(!acc.is_saturated());
 
         let batch =
