+++ vortex-array/src/aggregate_fn/fns/sum/constant.rs
@@ -127,8 +127,7 @@ mod tests {
         let array = ConstantArray::new(Scalar::null(DType::Primitive(PType::U32, Nullable)), 10)
             .into_array();
         let result = sum(&array, &mut array_session().create_execution_ctx())?;
-        // SQL `SUM`: an all-null constant has no valid values.
-        assert_eq!(result, Scalar::null(DType::Primitive(PType::U64, Nullable)));
+        assert_eq!(result, Scalar::primitive(0u64, Nullable));
         Ok(())
     }
 
@@ -152,7 +151,7 @@ mod tests {
     fn sum_constant_bool_null() -> VortexResult<()> {
         let array = ConstantArray::new(Scalar::null(DType::Bool(Nullable)), 10).into_array();
         let result = sum(&array, &mut array_session().create_execution_ctx())?;
-        assert_eq!(result, Scalar::null(DType::Primitive(PType::U64, Nullable)));
+        assert_eq!(result, Scalar::primitive(0u64, Nullable));
         Ok(())
     }
 
@@ -188,7 +187,11 @@ mod tests {
         let result = sum(&array, &mut array_session().create_execution_ctx())?;
         assert_eq!(
             result,
-            Scalar::null(DType::Decimal(DecimalDType::new(20, 2), Nullable))
+            Scalar::decimal(
+                DecimalValue::I256(i256::ZERO),
+                DecimalDType::new(20, 2),
+                Nullable
+            )
         );
         Ok(())
     }
