+++ vortex-array/src/aggregate_fn/fns/sum/mod.rs
@@ -11,7 +11,7 @@ use vortex_error::VortexExpect;
 use vortex_error::VortexResult;
 use vortex_error::vortex_bail;
 use vortex_error::vortex_err;
-use vortex_mask::Mask;
+use vortex_error::vortex_panic;
 use vortex_session::VortexSession;
 use vortex_session::registry::CachedId;
 
@@ -23,56 +23,34 @@ use crate::ArrayRef;
 use crate::Canonical;
 use crate::Columnar;
 use crate::ExecutionCtx;
-use crate::IntoArray;
 use crate::aggregate_fn::Accumulator;
 use crate::aggregate_fn::AggregateFnId;
 use crate::aggregate_fn::AggregateFnVTable;
 use crate::aggregate_fn::DynAccumulator;
 use crate::aggregate_fn::NumericalAggregateOpts;
-use crate::aggregate_fn::fns::total::SumState;
-use crate::aggregate_fn::fns::total::checked_add_i64;
-use crate::aggregate_fn::fns::total::checked_add_u64;
-use crate::aggregate_fn::fns::total::make_zero_state;
-use crate::aggregate_fn::fns::total::sum_decimal_dtype;
-use crate::arrays::Bool;
-use crate::arrays::BoolArray;
-use crate::arrays::Struct;
-use crate::arrays::StructArray;
-use crate::arrays::bool::BoolArrayExt;
-use crate::arrays::masked::mask_validity_canonical;
-use crate::arrays::struct_::StructArrayExt;
 use crate::dtype::DType;
-use crate::dtype::FieldName;
-use crate::dtype::FieldNames;
+use crate::dtype::DecimalDType;
+use crate::dtype::MAX_PRECISION;
 use crate::dtype::Nullability;
 use crate::dtype::PType;
-use crate::dtype::StructFields;
 use crate::expr::stats::Precision;
 use crate::expr::stats::Stat;
 use crate::expr::stats::StatsProvider;
 use crate::expr::stats::StatsProviderExt;
+use crate::scalar::DecimalValue;
 use crate::scalar::Scalar;
-use crate::validity::Validity;
 
-/// Return the SQL sum of an array: null when the array has no valid values or the sum
-/// overflows.
+/// The monoid sum of an array: zero when there are no valid values, null on overflow.
 ///
-/// See [`Sum`] for details.
+/// This is the value stored as `Stat::Sum` and merged across chunks, zones, and files: the
+/// empty sum is zero so that partials merge as a monoid (`combine(empty, x) = x`). For the
+/// SQL rule (a sum over zero valid values is null) use
+/// [`standard_sum`](crate::aggregate_fn::fns::standard_sum::standard_sum) /
+/// [`StandardSum`](crate::aggregate_fn::fns::standard_sum::StandardSum).
 pub fn sum(array: &ArrayRef, ctx: &mut ExecutionCtx) -> VortexResult<Scalar> {
-    // Short-circuit using cached array statistics. `Stat::Sum` is the monoid sum (zero for an
-    // array with no valid values), so the SQL empty-sum rule needs the null count: when it is
-    // unknown for a nullable array, fall through and compute rather than trust the cache.
+    // Short-circuit using cached array statistics.
     if let Precision::Exact(sum_scalar) = array.statistics().get(Stat::Sum) {
-        if !array.dtype().is_nullable() {
-            return Ok(sum_scalar);
-        }
-        match array.statistics().get_as::<u64>(Stat::NullCount) {
-            Precision::Exact(null_count) if null_count == array.len() as u64 => {
-                return Ok(Scalar::null(sum_scalar.dtype().as_nullable()));
-            }
-            Precision::Exact(_) => return Ok(sum_scalar),
-            _ => {}
-        }
+        return Ok(sum_scalar);
     }
 
     // Compute using Accumulator<Sum>.
@@ -85,8 +63,7 @@ pub fn sum(array: &ArrayRef, ctx: &mut ExecutionCtx) -> VortexResult<Scalar> {
     acc.accumulate(array, ctx)?;
     let result = acc.finish()?;
 
-    // Cache the computed sum as a statistic (only if non-null, i.e. no overflow and at least
-    // one valid value).
+    // Cache the computed sum as a statistic (only if non-null, i.e. no overflow).
     if let Some(val) = result.value().cloned() {
         array.statistics().set(Stat::Sum, Precision::Exact(val));
     }
@@ -94,31 +71,34 @@ pub fn sum(array: &ArrayRef, ctx: &mut ExecutionCtx) -> VortexResult<Scalar> {
     Ok(result)
 }
 
-/// Sum an array following SQL `SUM` semantics.
+/// The monoid sum aggregate: an all-invalid or empty sum is zero, an overflowing sum is null.
 ///
-/// A sum over zero valid values yields null (the SQL rule: nulls are eliminated, and the sum
-/// of an empty set is null), and a sum that overflows yields null. This is a distinct
-/// aggregate from the monoid [`Total`](crate::aggregate_fn::fns::total::Total), whose empty sum is
-/// zero and whose plain partial is the persisted form of sum statistics; `Sum`'s
-/// `{sum, seen}` partial is never written to statistics.
-///
-/// The partial state is a `{sum, seen}` struct: `sum` is the running monoid value (null once
-/// saturated by overflow, which poisons merges), and `seen` records whether any valid value
-/// contributed (merged with OR). `finalize` maps `seen == false` to null. `seen` is decided by
-/// validity alone: NaN values are valid, so with `skip_nans` a sum over only NaNs is `0`, not
-/// null.
+/// This aggregate's plain partial is the persisted wire form of sum statistics (zone maps,
+/// file stats), so its partial dtype and semantics must stay stable. The SQL `SUM` rule is
+/// provided by the separate [`StandardSum`](crate::aggregate_fn::fns::standard_sum::StandardSum)
+/// aggregate.
 ///
 /// NaN handling for float inputs is controlled by [`NumericalAggregateOpts`]: with `skip_nans` (the
 /// default) NaN values contribute nothing, otherwise any NaN value poisons the sum to NaN.
 #[derive(Clone, Debug)]
 pub struct Sum;
 
+// Both Spark and DataFusion use this heuristic.
+// - https://github.com/apache/spark/blob/fcf636d9eb8d645c24be3db2d599aba2d7e2955a/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/expressions/aggregate/Sum.scala#L66
+// - https://github.com/apache/datafusion/blob/4153adf2c0f6e317ef476febfdc834208bd46622/datafusion/functions-aggregate/src/sum.rs#L188
+pub(crate) fn sum_decimal_dtype(input: &DecimalDType) -> DecimalDType {
+    DecimalDType::new(
+        u8::min(MAX_PRECISION, input.precision() + 10),
+        input.scale(),
+    )
+}
+
 impl AggregateFnVTable for Sum {
     type Options = NumericalAggregateOpts;
     type Partial = SumPartial;
 
     fn id(&self) -> AggregateFnId {
-        static ID: CachedId = CachedId::new("vortex.sql_sum");
+        static ID: CachedId = CachedId::new("vortex.sum");
         *ID
     }
 
@@ -162,7 +142,7 @@ impl AggregateFnVTable for Sum {
     }
 
     fn partial_dtype(&self, options: &Self::Options, input_dtype: &DType) -> Option<DType> {
-        Some(sum_partial_dtype(self.return_dtype(options, input_dtype)?))
+        self.return_dtype(options, input_dtype)
     }
 
     fn empty_partial(
@@ -178,45 +158,19 @@ impl AggregateFnVTable for Sum {
         Ok(SumPartial {
             return_dtype,
             current: Some(initial),
-            seen: false,
             skip_nans: options.skip_nans,
         })
     }
 
     fn combine_partials(&self, partial: &mut Self::Partial, other: Scalar) -> VortexResult<()> {
-        // Partials are `{sum, seen}` structs. A plain (non-struct) scalar is the legacy or
-        // statistic form: a monoid sum value from an older writer or a cached `Stat::Sum`. It
-        // cannot distinguish an empty sum from a zero sum, so it is treated as seen.
-        let (sum_value, other_seen) = if matches!(other.dtype(), DType::Struct(..)) {
-            if other.is_null() {
-                // A null struct partial carries no recoverable state; treat as saturated.
-                partial.seen = true;
-                partial.current = None;
-                return Ok(());
-            }
-            let fields = other.as_struct();
-            let sum_value = fields
-                .field("sum")
-                .ok_or_else(|| vortex_err!("Sum partial is missing the `sum` field"))?;
-            let other_seen = fields
-                .field("seen")
-                .and_then(|seen| seen.as_bool().value())
-                .unwrap_or(true);
-            (sum_value, other_seen)
-        } else {
-            (other, true)
-        };
-
-        partial.seen |= other_seen;
-        if sum_value.is_null() {
-            // A null sum value means the sub-accumulator saturated (overflow).
+        if other.is_null() {
+            // A null partial means the sub-accumulator saturated (overflow).
             partial.current = None;
             return Ok(());
         }
         let Some(ref mut inner) = partial.current else {
             return Ok(());
         };
-        let other = sum_value;
         let saturated = match inner {
             SumState::Unsigned(acc) => {
                 let val = other
@@ -261,18 +215,23 @@ impl AggregateFnVTable for Sum {
     }
 
     fn to_scalar(&self, partial: &Self::Partial) -> VortexResult<Scalar> {
-        Ok(Scalar::struct_(
-            sum_partial_dtype(partial.return_dtype.as_nullable()),
-            vec![
-                sum_value_scalar(partial),
-                Scalar::bool(partial.seen, Nullability::NonNullable),
-            ],
-        ))
+        Ok(match &partial.current {
+            None => Scalar::null(partial.return_dtype.as_nullable()),
+            Some(SumState::Unsigned(v)) => Scalar::primitive(*v, Nullability::Nullable),
+            Some(SumState::Signed(v)) => Scalar::primitive(*v, Nullability::Nullable),
+            Some(SumState::Float(v)) => Scalar::primitive(*v, Nullability::Nullable),
+            Some(SumState::Decimal { value, .. }) => {
+                let decimal_dtype = *partial
+                    .return_dtype
+                    .as_decimal_opt()
+                    .vortex_expect("return dtype must be decimal");
+                Scalar::decimal(*value, decimal_dtype, Nullability::Nullable)
+            }
+        })
     }
 
     fn reset(&self, partial: &mut Self::Partial) {
         partial.current = Some(make_zero_state(&partial.return_dtype));
-        partial.seen = false;
     }
 
     #[inline]
@@ -290,26 +249,28 @@ impl AggregateFnVTable for Sum {
         batch: &ArrayRef,
         _ctx: &mut ExecutionCtx,
     ) -> VortexResult<bool> {
-        // `Stat::Sum` is the monoid (NaN-skipping) sum, so the default NaN-skipping path can
-        // consume it directly. `Sum` has no Stat slot of its own — the shortcut lives here
-        // rather than in the accumulator's stats bridge.
-        if partial.skip_nans {
-            return try_accumulate_cached_sum(self, partial, batch);
-        }
-        // NaN-including float sums need a NaN-free batch before the cached sum applies;
-        // everything else takes the default dispatch path.
-        if !matches!(partial.current, Some(SumState::Float(_))) {
+        // NaN-aware shortcircuits only apply to NaN-including float sums; everything else takes
+        // the default dispatch path.
+        if partial.skip_nans || !matches!(partial.current, Some(SumState::Float(_))) {
             return Ok(false);
         }
         match batch.statistics().get_as::<u64>(Stat::NaNCount) {
             Precision::Exact(0) => {
                 // NaN-free batch: the cached NaN-skipping sum (if any) equals the
                 // NaN-including sum.
-                try_accumulate_cached_sum(self, partial, batch)
+                if let Precision::Exact(sum) = batch.statistics().get(Stat::Sum) {
+                    let sum = if sum.dtype() == &partial.return_dtype {
+                        sum
+                    } else {
+                        sum.cast(&partial.return_dtype)?
+                    };
+                    self.combine_partials(partial, sum)?;
+                    return Ok(true);
+                }
+                Ok(false)
             }
             Precision::Exact(_) => {
-                // At least one NaN value (a valid value): the sum is NaN without scanning.
-                partial.seen = true;
+                // At least one NaN value: the sum is NaN without scanning the batch.
                 if let Some(SumState::Float(acc)) = partial.current.as_mut() {
                     *acc = f64::NAN;
                 }
@@ -327,9 +288,6 @@ impl AggregateFnVTable for Sum {
     ) -> VortexResult<()> {
         // Constants compute scalar * len and combine via combine_partials.
         if let Columnar::Constant(c) = batch {
-            // Any valid value counts as seen, including NaN and `false` constants that
-            // contribute nothing to the running sum.
-            partial.seen |= !c.scalar().is_null() && !c.is_empty();
             // NaN constants are treated as missing when skipping NaNs.
             if partial.skip_nans && c.scalar().as_primitive_opt().is_some_and(|p| p.is_nan()) {
                 return Ok(());
@@ -348,11 +306,9 @@ impl AggregateFnVTable for Sum {
 
         let result = match batch {
             Columnar::Canonical(c) => match c {
-                Canonical::Primitive(p) => {
-                    accumulate_primitive(&mut inner, p, ctx, skip_nans, &mut partial.seen)
-                }
-                Canonical::Bool(b) => accumulate_bool(&mut inner, b, ctx, &mut partial.seen),
-                Canonical::Decimal(d) => accumulate_decimal(&mut inner, d, ctx, &mut partial.seen),
+                Canonical::Primitive(p) => accumulate_primitive(&mut inner, p, ctx, skip_nans),
+                Canonical::Bool(b) => accumulate_bool(&mut inner, b, ctx),
+                Canonical::Decimal(d) => accumulate_decimal(&mut inner, d, ctx),
                 _ => vortex_bail!("Unsupported canonical type for sum: {}", batch.dtype()),
             },
             Columnar::Constant(_) => unreachable!(),
@@ -369,71 +325,13 @@ impl AggregateFnVTable for Sum {
         Ok(())
     }
 
-    fn finalize(&self, partials: ArrayRef, ctx: &mut ExecutionCtx) -> VortexResult<ArrayRef> {
-        // Entries that saw no valid values finalize to null (SQL `SUM`), while a null `sum`
-        // field (overflow) and null partial rows (e.g. null groups) stay null via the
-        // validity intersection. Implemented structurally rather than as a `get_item`/`mask`
-        // expression: expression construction and dispatch cost multiples of the whole
-        // aggregation at small group counts.
-        let len = partials.len();
-        let states = match partials.as_opt::<Struct>() {
-            Some(states) => states.into_owned(),
-            None => partials.execute::<StructArray>(ctx)?,
-        };
-        let struct_mask = states.as_ref().validity()?.execute_mask(len, ctx)?;
-        let seen = states.unmasked_field_by_name("seen")?.clone();
-        let seen = match seen.as_opt::<Bool>() {
-            Some(seen) => seen.into_owned(),
-            None => seen.execute::<BoolArray>(ctx)?,
-        };
-        let valid = &struct_mask & &Mask::from_buffer(seen.to_bit_buffer());
-
-        let sum = states.unmasked_field_by_name("sum")?.clone();
-        if valid.all_true() {
-            // Every partial row is a seen, valid group: the sums are already the result.
-            return Ok(sum);
-        }
-        let sum = sum.execute::<Canonical>(ctx)?;
-        Ok(
-            mask_validity_canonical(sum, Validity::from_mask(valid, Nullability::Nullable), ctx)?
-                .into_array(),
-        )
+    fn finalize(&self, partials: ArrayRef, _ctx: &mut ExecutionCtx) -> VortexResult<ArrayRef> {
+        Ok(partials)
     }
 
     fn finalize_scalar(&self, partial: &Self::Partial) -> VortexResult<Scalar> {
-        if !partial.seen {
-            return Ok(Scalar::null(partial.return_dtype.as_nullable()));
-        }
-        Ok(sum_value_scalar(partial))
+        self.to_scalar(partial)
     }
-}
-
-/// Consume a batch's cached monoid `Stat::Sum` instead of scanning it. The cached sum cannot
-/// carry `seen`, so the batch's null count decides between the identity (all null) and a seen
-/// contribution; when either statistic is missing the caller falls through to a real scan.
-fn try_accumulate_cached_sum(
-    vtable: &Sum,
-    partial: &mut SumPartial,
-    batch: &ArrayRef,
-) -> VortexResult<bool> {
-    let Precision::Exact(sum) = batch.statistics().get(Stat::Sum) else {
-        return Ok(false);
-    };
-    match batch.statistics().get_as::<u64>(Stat::NullCount) {
-        Precision::Exact(null_count) if null_count == batch.len() as u64 => {
-            // No valid values: the batch is the identity.
-            return Ok(true);
-        }
-        Precision::Exact(_) => partial.seen = true,
-        _ => return Ok(false),
-    }
-    let sum = if sum.dtype() == &partial.return_dtype {
-        sum
-    } else {
-        sum.cast(&partial.return_dtype)?
-    };
-    vtable.combine_partials(partial, sum)?;
-    Ok(true)
 }
 
 /// The group state for a sum aggregate, containing the accumulated value and configuration
@@ -442,42 +340,59 @@ pub struct SumPartial {
     return_dtype: DType,
     /// The current accumulated state, or `None` if saturated (checked overflow).
     current: Option<SumState>,
-    /// Whether at least one valid value has been accumulated. A sum over zero valid values
-    /// finalizes to null (SQL `SUM` semantics) rather than the monoid zero.
-    seen: bool,
     /// Whether NaN values in float inputs are skipped.
     skip_nans: bool,
 }
 
-/// The partial dtype for a sum whose result is `sum_dtype`: a `{sum, seen}` struct, where
-/// `sum` is the running monoid value (null once saturated by overflow, poisoning merges) and
-/// `seen` records whether any valid value contributed (merged with OR). Keeping the flag
-/// separate from the sum lets merges stay a monoid — the identity is `{0, false}` — while
-/// `finalize` maps unseen sums to null.
-fn sum_partial_dtype(sum_dtype: DType) -> DType {
-    DType::Struct(
-        StructFields::new(
-            FieldNames::from_iter([FieldName::from("sum"), FieldName::from("seen")]),
-            vec![sum_dtype, DType::Bool(Nullability::NonNullable)],
-        ),
-        Nullability::Nullable,
-    )
+/// The accumulated sum value.
+// TODO(ngates): instead of an enum, we should use a Box<dyn State> to avoid dispatcher over the
+//  input type every time? Perhaps?
+pub enum SumState {
+    Unsigned(u64),
+    Signed(i64),
+    Float(f64),
+    Decimal {
+        value: DecimalValue,
+        dtype: DecimalDType,
+    },
 }
 
-/// The running sum as a nullable scalar of the return dtype (null when saturated by overflow).
-fn sum_value_scalar(partial: &SumPartial) -> Scalar {
-    match &partial.current {
-        None => Scalar::null(partial.return_dtype.as_nullable()),
-        Some(SumState::Unsigned(v)) => Scalar::primitive(*v, Nullability::Nullable),
-        Some(SumState::Signed(v)) => Scalar::primitive(*v, Nullability::Nullable),
-        Some(SumState::Float(v)) => Scalar::primitive(*v, Nullability::Nullable),
-        Some(SumState::Decimal { value, .. }) => {
-            let decimal_dtype = *partial
-                .return_dtype
-                .as_decimal_opt()
-                .vortex_expect("return dtype must be decimal");
-            Scalar::decimal(*value, decimal_dtype, Nullability::Nullable)
+pub(crate) fn make_zero_state(return_dtype: &DType) -> SumState {
+    match return_dtype {
+        DType::Primitive(ptype, _) => match ptype {
+            PType::U8 | PType::U16 | PType::U32 | PType::U64 => SumState::Unsigned(0),
+            PType::I8 | PType::I16 | PType::I32 | PType::I64 => SumState::Signed(0),
+            PType::F16 | PType::F32 | PType::F64 => SumState::Float(0.0),
+        },
+        DType::Decimal(decimal, _) => SumState::Decimal {
+            value: DecimalValue::zero(decimal),
+            dtype: *decimal,
+        },
+        _ => vortex_panic!("Unsupported sum type"),
+    }
+}
+
+/// Checked add for u64, returning true if overflow occurred.
+#[inline(always)]
+pub(crate) fn checked_add_u64(acc: &mut u64, val: u64) -> bool {
+    match acc.checked_add(val) {
+        Some(r) => {
+            *acc = r;
+            false
         }
+        None => true,
+    }
+}
+
+/// Checked add for i64, returning true if overflow occurred.
+#[inline(always)]
+pub(crate) fn checked_add_i64(acc: &mut i64, val: i64) -> bool {
+    match acc.checked_add(val) {
+        Some(r) => {
+            *acc = r;
+            false
+        }
+        None => true,
     }
 }
 
@@ -499,7 +414,6 @@ mod tests {
     use crate::aggregate_fn::NumericalAggregateOpts;
     use crate::aggregate_fn::fns::sum::Sum;
     use crate::aggregate_fn::fns::sum::sum;
-    use crate::aggregate_fn::fns::total::total;
     use crate::array_session;
     use crate::arrays::BoolArray;
     use crate::arrays::ChunkedArray;
@@ -617,100 +531,6 @@ mod tests {
     // State merge tests (vtable-level)
 
     #[test]
-    fn sum_state_empty_is_null() -> VortexResult<()> {
-        // A state that never saw a valid value finalizes to null, and combining empty states
-        // stays empty.
-        let dtype = DType::Primitive(PType::I32, Nullability::NonNullable);
-        let mut state = Sum.empty_partial(&NumericalAggregateOpts::default(), &dtype)?;
-        let empty = Sum.to_scalar(&state)?;
-        Sum.combine_partials(&mut state, empty)?;
-        assert!(Sum.finalize_scalar(&state)?.is_null());
-        Ok(())
-    }
-
-    #[test]
-    fn sum_state_empty_is_identity() -> VortexResult<()> {
-        // Combining an empty state into a seen state changes nothing: `{0, false}` is the
-        // identity of the `{sum, seen}` monoid.
-        let dtype = DType::Primitive(PType::I32, Nullability::NonNullable);
-        let mut state = Sum.empty_partial(&NumericalAggregateOpts::default(), &dtype)?;
-        Sum.combine_partials(&mut state, Scalar::primitive(100i64, Nullable))?;
-
-        let empty =
-            Sum.to_scalar(&Sum.empty_partial(&NumericalAggregateOpts::default(), &dtype)?)?;
-        Sum.combine_partials(&mut state, empty)?;
-
-        let result = Sum.finalize_scalar(&state)?;
-        assert_eq!(result.as_primitive().typed_value::<i64>(), Some(100));
-        Ok(())
-    }
-
-    #[test]
-    fn sum_state_overflow_poisons_but_stays_seen() -> VortexResult<()> {
-        // Overflow (a null `sum` field) poisons the merge even when combined with later
-        // values: the result is null via the sum value, not via `seen`.
-        let dtype = DType::Primitive(PType::I64, Nullability::NonNullable);
-        let mut overflowed = Sum.empty_partial(&NumericalAggregateOpts::default(), &dtype)?;
-        Sum.combine_partials(&mut overflowed, Scalar::primitive(i64::MAX, Nullable))?;
-        Sum.combine_partials(&mut overflowed, Scalar::primitive(1i64, Nullable))?;
-        let overflowed = Sum.to_scalar(&overflowed)?;
-
-        let mut state = Sum.empty_partial(&NumericalAggregateOpts::default(), &dtype)?;
-        Sum.combine_partials(&mut state, Scalar::primitive(5i64, Nullable))?;
-        Sum.combine_partials(&mut state, overflowed)?;
-        Sum.combine_partials(&mut state, Scalar::primitive(7i64, Nullable))?;
-
-        assert!(Sum.finalize_scalar(&state)?.is_null());
-        Ok(())
-    }
-
-    #[test]
-    fn sum_all_nan_is_zero_not_null() -> VortexResult<()> {
-        // NaNs are valid values: with the default `skip_nans` they contribute nothing, but
-        // the sum is a genuine `0.0`, unlike an all-null array whose sum is null.
-        let arr =
-            PrimitiveArray::new(buffer![f64::NAN, f64::NAN], Validity::NonNullable).into_array();
-        let result = sum(&arr, &mut array_session().create_execution_ctx())?;
-        assert_eq!(result.as_primitive().typed_value::<f64>(), Some(0.0));
-        Ok(())
-    }
-
-    #[test]
-    fn total_is_monoid_while_sum_is_sql() -> VortexResult<()> {
-        // The persisted statistic keeps the monoid semantics (zero for all-null) that zone
-        // and chunk merging require, while `sum` applies the SQL rule.
-        let mut ctx = array_session().create_execution_ctx();
-        let arr = PrimitiveArray::from_option_iter([None::<i32>, None, None]).into_array();
-        assert_eq!(
-            total(&arr, &mut ctx)?.as_primitive().typed_value::<i64>(),
-            Some(0)
-        );
-        // The cached monoid statistic must not leak through `sum`'s cache short-circuit.
-        assert!(sum(&arr, &mut ctx)?.is_null());
-        Ok(())
-    }
-
-    #[test]
-    fn grouped_sum_fallback_empty_and_all_null_groups() -> VortexResult<()> {
-        // Bool elements are rejected by the primitive grouped kernel, forcing the generic
-        // per-group fallback: empty and all-null groups have null sums there too.
-        let mut ctx = array_session().create_execution_ctx();
-        let elements = BoolArray::from_iter([Some(true), Some(true), None, None]).into_array();
-        let groups = ListViewArray::try_new(
-            elements,
-            buffer![0i32, 2, 2].into_array(),
-            buffer![2i32, 0, 2].into_array(),
-            Validity::NonNullable,
-        )?
-        .into_array();
-
-        let result = run_grouped_sum(&groups, &DType::Bool(Nullable))?;
-        let expected = PrimitiveArray::from_option_iter([Some(2u64), None, None]).into_array();
-        assert_arrays_eq!(&result, &expected, &mut ctx);
-        Ok(())
-    }
-
-    #[test]
     fn sum_state_merge() -> VortexResult<()> {
         let dtype = DType::Primitive(PType::I32, Nullability::NonNullable);
         let mut state = Sum.empty_partial(&NumericalAggregateOpts::default(), &dtype)?;
@@ -721,7 +541,7 @@ mod tests {
         let scalar2 = Scalar::primitive(50i64, Nullable);
         Sum.combine_partials(&mut state, scalar2)?;
 
-        let result = Sum.finalize_scalar(&state)?;
+        let result = Sum.to_scalar(&state)?;
         Sum.reset(&mut state);
         assert_eq!(result.as_primitive().typed_value::<i64>(), Some(150));
         Ok(())
@@ -835,8 +655,7 @@ mod tests {
         let elem_dtype = DType::Primitive(PType::I32, Nullable);
         let result = run_grouped_sum(&groups.into_array(), &elem_dtype)?;
 
-        // The all-null group has a null sum (SQL `SUM` semantics).
-        let expected = PrimitiveArray::from_option_iter([None, Some(7i64)]).into_array();
+        let expected = PrimitiveArray::from_option_iter([Some(0i64), Some(7i64)]).into_array();
         assert_arrays_eq!(&result, &expected, &mut ctx);
         Ok(())
     }
@@ -929,7 +748,7 @@ mod tests {
     }
 
     #[test]
-    fn sum_chunked_floats_all_nulls_is_null() -> VortexResult<()> {
+    fn sum_chunked_floats_all_nulls_is_zero() -> VortexResult<()> {
         let chunk1 = PrimitiveArray::from_option_iter::<f32, _>(vec![None, None, None]);
         let chunk2 = PrimitiveArray::from_option_iter::<f32, _>(vec![None, None]);
         let dtype = chunk1.dtype().clone();
@@ -938,8 +757,7 @@ mod tests {
             &chunked.into_array(),
             &mut array_session().create_execution_ctx(),
         )?;
-        // SQL `SUM`: no valid values across any chunk yields null.
-        assert!(result.is_null());
+        assert_eq!(result, Scalar::primitive(0f64, Nullable));
         Ok(())
     }
